Can now recognize if the Opal Kelly FPGA board is usb2 or usb3.
For usb3, will use block throttled pipeouts to read out data from the
FPGA, this is because usb3 is fast and block throttle is needed to
ensure data integrity.

The "blockSize" parameter for the ReadFromBlockPipeOut() function is
defined by BTblockSize in globalconstants.h. Default ot 512 bytes (must
be powers of 2.

Rhd2000evalboard is now initialized with a blockSize as a parameter,
ignored if usb2. Its "open()" method now maintains the board type
parameter, and is passed to mainwindow class as well.

Because of the block throttling, the "length" parameter for the
ReadFromBlockPipeOut() function must be divisible by blockSize.
Previously, initialization sequences sets maxtimestp to 60 (as the
length of the longest auxcommandList), but the resulting data will not
have a size divisible by blockSize. Now it's set to 64 (any greater
powers of 2 should be fine).

rhd2000datablock now takes boolean "usb3" as a parameter - the default
SAMPLES_PER_DATA_BLOCK parameter is now 60 or 64 depending on its value.

Note that the blockSize parameter set here must be 4x the BLOCK_SIZE
parameter in FPGA's HDL code (SDRAM_FIFO.v). This is because the
opal kelly usb3 pipeout is 32-bits (or 4 bytes) wide.

With the current configuration:

- HDL:
  - BLOCK_SIZE = 128 (SDRAM_FIFO)
  - BURST_LEN = 2    (ddr2_state_machine)
  - FIFO_SIZE = 1024 (ddr2_state_machine): must match the depth of the
    input and output FIFO modules.
- C++:
  - BTBlockSize = 512 (globalconstants.h)
  - SAMPLES_PER_DATA_BLOCK_USB3 = 64 or 128 (rhd2000datablock.h)

With 1-5, and 8 datastreams enabled, no data corruption happens.
But with 6 or 7 datastreams enabled, ReadFromBlockPipeOut() returns the
PipeRead failed error...need to investigate.